move: event prize section component
(cherry picked from commit 5320c01bbccc3d420a98a8b41d89c9792bea40f1)
