+++ packages/user/src/components/home/eventPrizes/PrizeCard.tsx
