+++ packages/user/src/components/home/index.ts
@@ -1,2 +1,2 @@
 export { default as EventGuidelines } from './EventGuidelines';
-export { default as EventPrizes } from './EventPrizes';
+export { default as EventPrizes } from './eventPrizes';
